Initial efforts of porting to VFL
This commit begins the (non-trivial) process of porting all of
the vbnmr functionality over to using the vfl library.

The two projects have progressed essentially in parallel,
with vbnmr being the initial design, and vfl arising from
the realization that the mathematics are highly general.

This port will greatly simplify vbnmr, as it removes a ton
of code that is already provided by vfl. Further, it makes
maintenance and extension of vbnmr easier, as the underlying
vfl framework is highly modular.

So in short, this commit introduces the meat and potatoes of
vbnmr: the quadrature factor and the hybrid fixed-tau vfr plus
gaussian process model.

More work in both source trees will be required to fully port
vbnmr into vfl, but this is a definite step in that direction.