Added a test of mean-field for 2D NUS data
This commit demonstrates the use of quad_meanfield() for
two-dimensional data, nonuniformly sampled from an underlying
uniform grid.

It also removes the natural gradient toy2d test, which could
not match mean-field performance. In fact, just one iteration
of natural gradient ascent seemed to completely negate the
results of an initial mean-field fit.

Of course, this difficulty with natural gradients is not
wholly unexpected, due to the strong multi-modality of
the objective.