Fixed a bug in the quadrature factor
This commit fixes a bug that was in quad_set_dims(),
where P and K were swapped in the call to factor_resize().

While this fixes one bug, there is still another that
has proved highly difficult to track down. Something
is causing a segfault after resizing a quad factor
and then accessing an element of its mean list or
precision list.